change userid double to int
changed userid type from int to double because the after the hashing some users couldn't be added anymore because their number was higher than int can be

+++ setup.sql
@@ -13,6 +13,6 @@ CREATE TABLE IF NOT EXISTS `messagetable` (
 );
 
 CREATE TABLE IF NOT EXISTS `optintable` (
-        `userid` INT NOT NULL,
+        `userid` DOUBLE NOT NULL,
         PRIMARY KEY (`userid`)
 );
